move object utils to localstack.utils.objects
